requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
pillow
python-multipart==0.0.12